requirements.txt
jinja2
pyyaml
pydantic
setuptools
datacontract-cli[databricks]
pyspark[sql,connect]
numpy
pandas
fastavro
